API v27

+++ libraries/WaspAES/WaspAES.cpp
@@ -1,5 +1,5 @@
 /*
- *  Copyright (C) 2016 Libelium Comunicaciones Distribuidas S.L.
+ *  Copyright (C) 2017 Libelium Comunicaciones Distribuidas S.L.
  *  http://www.libelium.com
  *
  *  This program is free software: you can redistribute it and/or modify
@@ -15,7 +15,7 @@
  *  You should have received a copy of the GNU Lesser General Public License
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
- *  Version:		3.0
+ *  Version:		3.1
  *  Design:			David Gascón
  *  Implementation:	Alvaro Gonzalez, Yuri Carmona
  *
@@ -359,7 +359,7 @@ void WaspAES::assignBlock(uint8_t *a, uint8_t *b)
  * to the original message in 'message'.
  *
  */
-uint8_t WaspAES::sizeOfBlocks(char* message)
+uint16_t WaspAES::sizeOfBlocks(char* message)
 {
 	uint16_t length;	
 	length = strlen(message);
@@ -374,7 +374,7 @@ uint8_t WaspAES::sizeOfBlocks(char* message)
  * to the original message in 'message'.
  *
  */
-uint8_t WaspAES::sizeOfBlocks(uint16_t length)
+uint16_t WaspAES::sizeOfBlocks(uint16_t length)
 {
 	float aux;
